+++ src/constraints/not_equal.jl
@@ -106,3 +106,26 @@ function is_constraint_solved(
     coeffs = [t.coefficient for t in fct.terms]
     return get_approx_discrete(sum(values .* coeffs) + fct.constant) != set.value
 end
+
+"""
+    is_constraint_violated(
+        com::CoM,
+        constraint::LinearConstraint,
+        fct::SAF{T},
+        set::NotEqualTo{T},
+    ) where {T<:Real}
+
+Checks if the constraint is violated as it is currently set. This can happen inside an
+inactive reified or indicator constraint.
+"""
+function is_constraint_violated(
+    com::CoM,
+    constraint::LinearConstraint,
+    fct::SAF{T},
+    set::NotEqualTo{T},
+) where {T<:Real}
+    if all(isfixed(var) for var in com.search_space[constraint.indices])
+        return !is_constraint_solved(constraint, fct, set, [CS.value(var) for var in com.search_space[constraint.indices]])
+    end
+    return false
+end