+++ test/general.jl
@@ -139,9 +139,9 @@ function is_solved(com::CS.CoM)
     end
     for constraint in com.constraints
         values = CS.value.(com.search_space[constraint.indices])
-        c_solved =
-            CS.is_constraint_solved(constraint, constraint.fct, constraint.set, values)
-        if !c_solved
+        c_solved = CS.is_constraint_solved(constraint, constraint.fct, constraint.set, values)
+        c_violated = CS.is_constraint_violated(com, constraint, constraint.fct, constraint.set)
+        if !c_solved || c_violated
             @error "Constraint $(constraint.idx) is not solved"
             @error "Info about constraint: $(typeof(constraint)), $(typeof(constraint.fct)), $(typeof(constraint.set))"
             return false