+++ test/unit/constraints/alldifferent.jl
@@ -140,3 +140,31 @@ end
         @test sort(CS.values(com.search_space[ind])) == [3, 7]
     end
 end
+
+@testset "all different is_constraint_violated test" begin
+    m = Model(optimizer_with_attributes(CS.Optimizer, "no_prune" => true, "logging" => []))
+    @variable(m, -5 <= x[1:10] <= 5, Int)
+    @constraint(m, x in CS.AllDifferentSet())
+    optimize!(m)
+    com = JuMP.backend(m).optimizer.model.inner
+
+    constraint = get_constraints_by_type(com, CS.AllDifferentConstraint)[1]
+
+    variables = com.search_space
+    @test CS.fix!(com, variables[1], 1; check_feasibility = false)
+    @test CS.fix!(com, variables[2], 1; check_feasibility = false)
+    @test CS.is_constraint_violated(com, constraint, constraint.fct, constraint.set)
+
+    m = Model(optimizer_with_attributes(CS.Optimizer, "no_prune" => true, "logging" => []))
+    @variable(m, -5 <= x[1:10] <= 5, Int)
+    @constraint(m, x in CS.AllDifferentSet())
+    optimize!(m)
+    com = JuMP.backend(m).optimizer.model.inner
+
+    constraint = get_constraints_by_type(com, CS.AllDifferentConstraint)[1]
+
+    variables = com.search_space
+    @test CS.fix!(com, variables[1], 1; check_feasibility = false)
+    @test CS.fix!(com, variables[2], 2; check_feasibility = false)
+    @test !CS.is_constraint_violated(com, constraint, constraint.fct, constraint.set)
+end