+++ test/unit/constraints/eq_sum.jl
@@ -69,3 +69,31 @@
     end
     CS.values(com.search_space[constr_indices[1]]) == [2]
 end
+
+@testset "eqsum is_constraint_violated test" begin
+    m = Model(optimizer_with_attributes(CS.Optimizer, "no_prune" => true, "logging" => []))
+    @variable(m, -5 <= x[1:2] <= 5, Int)
+    @constraint(m, sum(x) == 10)
+    optimize!(m)
+    com = JuMP.backend(m).optimizer.model.inner
+
+    constraint = get_constraints_by_type(com, CS.LinearConstraint)[1]
+
+    variables = com.search_space
+    @test CS.fix!(com, variables[1], 1; check_feasibility = false)
+    @test CS.fix!(com, variables[2], 1; check_feasibility = false)
+    @test CS.is_constraint_violated(com, constraint, constraint.fct, constraint.set)
+
+    m = Model(optimizer_with_attributes(CS.Optimizer, "no_prune" => true, "logging" => []))
+    @variable(m, -5 <= x[1:2] <= 5, Int)
+    @constraint(m, sum(x) == 10)
+    optimize!(m)
+    com = JuMP.backend(m).optimizer.model.inner
+
+    constraint = get_constraints_by_type(com, CS.LinearConstraint)[1]
+
+    variables = com.search_space
+    @test CS.fix!(com, variables[1], 5; check_feasibility = false)
+    @test CS.fix!(com, variables[2], 5; check_feasibility = false)
+    @test !CS.is_constraint_violated(com, constraint, constraint.fct, constraint.set)
+end