+++ test/unit/constraints/geqset.jl
@@ -95,3 +95,31 @@
     @test sort(CS.values(com.search_space[constr_indices[2]])) == -5:4
     @test sort(CS.values(com.search_space[constr_indices[3]])) == 4:4
 end
+
+@testset "geqset is_constraint_violated test" begin
+    m = Model(optimizer_with_attributes(CS.Optimizer, "no_prune" => true, "logging" => []))
+    @variable(m, -5 <= x[1:5] <= 5, Int)
+    @constraint(m, x in CS.GeqSet())
+    optimize!(m)
+    com = JuMP.backend(m).optimizer.model.inner
+
+    constraint = com.constraints[1]
+
+    variables = com.search_space
+    @test CS.fix!(com, variables[1], 3; check_feasibility = false)
+    @test CS.remove_below!(com, variables[2], 4; check_feasibility = false)
+    @test CS.is_constraint_violated(com, constraint, constraint.fct, constraint.set)
+
+    m = Model(optimizer_with_attributes(CS.Optimizer, "no_prune" => true, "logging" => []))
+    @variable(m, -5 <= x[1:5] <= 5, Int)
+    @constraint(m, x in CS.GeqSet())
+    optimize!(m)
+    com = JuMP.backend(m).optimizer.model.inner
+
+    constraint = com.constraints[1]
+
+    variables = com.search_space
+    @test CS.fix!(com, variables[1], -5; check_feasibility = false)
+    @test CS.fix!(com, variables[2], -5; check_feasibility = false)
+    @test !CS.is_constraint_violated(com, constraint, constraint.fct, constraint.set)
+end